changed size_t to int

+++ shm/inc/Ship.hpp
@@ -5,13 +5,13 @@
 class Ship {
 public:
     Ship() = default;
-    Ship(size_t, const std::string&, size_t, size_t, size_t);
-    Ship(size_t, size_t, size_t);
+    Ship(int, const std::string&, size_t, size_t, size_t);
+    Ship(int, size_t, size_t);
 
     Ship& operator+=(const size_t);
     Ship& operator-=(const size_t);
 
-    size_t getId() const { return id_; };
+    int getId() const { return id_; };
     std::string getName() const { return name_; };
     size_t getSpeed() const { return speed_; };
     size_t getMaxCrew() const { return maxCrew_; };
